+++ Changelog.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## master
+
+Nothing here yet.
+
 ## v4.3.1 (2019-05-03)
 
 - Bugfix: Update 2019 Labour Day Holidays for China as changed by government recently (2019-03-22), by @iamsk, and thanks to @ltyely for their patch (#345 & #347).