+++ Changelog.md
@@ -4,7 +4,7 @@
 
 - Added California specific calendars: California Education, Berkeley, San Francisco, West Hollywood (#215).
 - Added a few refactors and tests for Australia Capital Territory holiday named "Family & Community Day", that lasted from 2007 to 2017 (#25).
-
+- Added South African 2019 National Elections as holiday
 ## v4.3.1 (2019-05-03)
 
 - Bugfix: Update 2019 Labour Day Holidays for China as changed by government recently (2019-03-22), by @iamsk, and thanks to @ltyely for their patch (#345 & #347).