+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Added California specific calendars: California Education, Berkeley, San Francisco, West Hollywood (#215).
 - Added a few refactors and tests for Australia Capital Territory holiday named "Family & Community Day", that lasted from 2007 to 2017 (#25).
+- Added South African 2019 National Elections as holiday (#350), by @RichardOB.
 
 ## v4.3.1 (2019-05-03)